closed #69 - anonyms store passes only locally

+++ frontend/src/entry-command.js
@@ -1,8 +1,9 @@
 import assert from 'assert';
 import { Command, execute, can_execute } from 'command-decorator'; // eslint-disable-line no-unused-vars
 import { http, parse_location } from 'src/plugins/http.js';
-import * as i18n from 'src/plugins/auth.js';
 import { notifier_error } from 'src/services/loader.js';
+import * as i18n from 'src/plugins/i18n.js';
+import * as auth from 'src/services/auth.js';
 
 // TODO: move it somewhere
 const API_ENTRIES_URL = 'api/entries';
@@ -31,11 +32,13 @@ export class EntryCommand extends Command {
         };
 
         let response;
-        if (this.entry.id !== undefined) {
-            response = await http.put(`${API_ENTRIES_URL}/${this.entry.id}`, dto);
-        } else {
-            response = await http.post(API_ENTRIES_URL, dto);
-            this.entry.id = parseInt(parse_location(response));
+        if (auth.is_authenticated()) {
+            if (this.entry.id !== undefined) {
+                response = await http.put(`${API_ENTRIES_URL}/${this.entry.id}`, dto);
+            } else {
+                response = await http.post(API_ENTRIES_URL, dto);
+                this.entry.id = parseInt(parse_location(response));
+            }
         }
         this.entry.user = dto.user;
         this.entry.title = dto.title;
@@ -59,8 +62,10 @@ export class EntryCommand extends Command {
     @notifier_error(i18n.terror)
     @execute
     async delete() {
-        if (this.entry.id !== undefined) {
-            await http.delete(`${API_ENTRIES_URL}/${this.entry.id}`);
+        if (auth.is_authenticated()) {
+            if (this.entry.id !== undefined) {
+                await http.delete(`${API_ENTRIES_URL}/${this.entry.id}`);
+            }
         }
         (this.ondelete || function() { })();
     }
